fixed sims tests
