+++ x/posts/types/models/post_test.go
@@ -335,37 +335,6 @@ func TestPost_Validate(t *testing.T) {
 			expError: "post subspace must be a valid sha-256 hash",
 		},
 		{
-			name: "Post creation data in future",
-			post: models.Post{
-				PostID:         id,
-				ParentID:       id2,
-				Message:        "Message",
-				AllowsComments: true,
-				Subspace:       "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e",
-				OptionalData:   map[string]string{},
-				Created:        time.Now().UTC().Add(time.Hour),
-				Creator:        owner,
-				Medias:         medias,
-			},
-			expError: "post creation date cannot be in the future",
-		},
-		{
-			name: "Post last edit date in future",
-			post: models.Post{
-				PostID:         id,
-				ParentID:       id2,
-				Message:        "Message",
-				AllowsComments: true,
-				Subspace:       "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e",
-				OptionalData:   map[string]string{},
-				Created:        time.Now().UTC(),
-				LastEdited:     time.Now().UTC().Add(time.Hour),
-				Creator:        owner,
-				Medias:         medias,
-			},
-			expError: "post last edit date cannot be in the future",
-		},
-		{
 			name:     "Valid post without poll data",
 			post:     models.NewPost(id, "", "Message", true, "4e188d9c17150037d5199bbdb91ae1eb2a78a15aca04cb35530cccb81494b36e", map[string]string{}, date, owner).WithMedias(medias),
 			expError: "",
